refactor(gda): move project, export, meta groups into gda.commands (ADR-0040)
export.py absorbs export_run.py (incl. EXPORT_GET/RUN command homes);
meta.py owns info/skill/schema and absorbs skill_ops.py, closing over
the root app for the surface walk; gda.models no longer imports
gda.skill_targets. Surface manifest verified identical except one
`export run` help line, whose `gda.export_run.run_export_operation`
cross-reference now names the module that exists.
