+++ src/gda/models.py
@@ -19,7 +19,6 @@ from pydantic import (
 )
 
 from gda.execution import ExecutionKind
-from gda.skill_targets import SkillProvider, SkillScope, resolve_skill_dir
 
 
 class ErrorCategory(str, Enum):
@@ -93,25 +92,6 @@ class OperationErrorEnvelope(BaseModel):
     model_config = ConfigDict(extra="forbid")
 
     error: OperationError
-
-
-class InfoParams(BaseModel):
-    """The operation params of ``gda info`` — none (ADR-0004).
-
-    ``gda info`` takes no operation params, so its ``input`` schema is trivially
-    empty; this is expected, not an error. The model still exists so the
-    ``--schema`` document is derived model-side rather than hand-written.
-    """
-
-
-class SchemaAllParams(BaseModel):
-    """The operation params of ``gda schema`` — none (ADR-0012).
-
-    Like ``gda info``, ``gda schema`` takes no operation params, so its
-    ``input`` schema is trivially empty. The model still exists so ``gda schema
-    --schema`` is derived model-side rather than hand-written, keeping the meta
-    command self-describing under the same ADR-0004 gate as every other command.
-    """
 
 
 class LiveStackConstraints(BaseModel):
@@ -294,26 +274,6 @@ def normalize_path(path: str) -> str:
 NormalizedPath = Annotated[str, AfterValidator(normalize_path)]
 
 
-def normalize_export_output_path(path: str) -> str:
-    """Normalize an ``export run --output`` artifact path (#403).
-
-    Export runs the native Godot export with cwd set to the project directory.
-    A relative ``--output`` must therefore be made absolute against the invoker's
-    cwd before the runner sees it, or Godot writes into the project tree while
-    the result echoes an unlocatable relative string. Virtual paths keep the
-    shared path convention and pass through unchanged.
-    """
-    if "://" in path:
-        return path
-    expanded = Path(path).expanduser()
-    if expanded.is_absolute():
-        return str(expanded)
-    return str(Path.cwd() / expanded)
-
-
-ExportOutputPath = Annotated[str, AfterValidator(normalize_export_output_path)]
-
-
 # The one read-side value projection every value gda emits goes through
 # (ADR-0035): the shared field description for the dynamically-shaped `value`
 # fields. The field itself stays `Any` — a value's shape is not statically
@@ -442,219 +402,6 @@ class NodeProperty(BaseModel):
     )
 
 
-# --- project static-analysis reads (issue #116) -----------------------------
-#
-# Four read-only, project-wide analysis commands, all backed by a single static
-# project scan: find-references, dependencies, find-unused-resources, statistics.
-# The scan parses files as TEXT (.tscn/.tres ext_resource paths, .gd
-# preload/load/class_name references) — it never instantiates a scene or loads a
-# script, so it honors the read trust boundary (issue #30). The only residual
-# project-code execution is the engine constructing the project's autoloads at
-# startup, inherent to any ``--project`` op (issue #61), documented on the params.
-
-
-class ProjectFindReferencesParams(BaseModel):
-    """The operation params of ``gda project find-references`` (issue #116).
-
-    ``target`` is what to find references TO: a resource's ``res://`` path (a
-    scene, script, image, ``.tres`` resource — anything addressable by path), or
-    a script ``class_name``. The scan reads project files as TEXT and never
-    instantiates a scene or loads a script (issue #30); the only project code that
-    runs is the project's autoloads, constructed by the engine at startup on any
-    ``--project`` op (issue #61).
-    """
-
-    target: NormalizedPath = Field(
-        description=(
-            "What to find references to: a resource's res:// path (scene, script, "
-            "image, .tres, …) or a script class_name."
-        )
-    )
-
-
-class ExportListParams(BaseModel):
-    """The operation params of ``gda export list`` — none (ADR-0004).
-
-    ``export list`` enumerates the export presets defined in the resolved
-    project's ``export_presets.cfg``; the project is process context
-    (``--project``), not an operation param (ADR-0006), so the ``input`` schema
-    is trivially empty, exactly like ``scene list`` / ``script list``.
-    """
-
-
-class ListedPreset(BaseModel):
-    """One enumerated export preset of ``gda export list`` (issue #114).
-
-    Read cheaply from ``export_presets.cfg`` (a ``ConfigFile`` parse, no engine
-    export run): ``name`` is the preset's display name — the address an agent
-    feeds back into ``gda export get`` — ``platform`` the target platform (e.g.
-    ``Linux/X11``, ``Web``), and ``runnable`` whether the preset is marked
-    runnable (one-click deploy). ``index`` is the preset's 0-based position in
-    the file (its ``preset.N`` section number), stable across a single read.
-    """
-
-    index: int = Field(
-        description="The preset's 0-based position in export_presets.cfg (its preset.N section number)."
-    )
-    name: str = Field(description="The preset's display name.")
-    platform: str = Field(
-        description="The preset's target platform (e.g. Linux/X11, Web, macOS)."
-    )
-    runnable: bool = Field(
-        description="Whether the preset is marked runnable (one-click deploy)."
-    )
-
-
-class ExportListResult(BaseModel):
-    """The result of ``gda export list``: the project's enumerated export presets.
-
-    A project whose ``export_presets.cfg`` exists but defines no presets is a
-    valid, empty listing — ``presets == []`` — not a failure. A project with no
-    ``export_presets.cfg`` at all is the ``export_presets_not_found`` failure
-    (it has no export configuration), distinct from an empty one.
-    """
-
-    presets: list[ListedPreset]
-
-
-class ExportGetParams(BaseModel):
-    """The operation params of ``gda export get`` (issue #114).
-
-    ``preset`` addresses an export preset by its display name (as ``export
-    list`` reports it). An unknown name is the ``export_preset_not_found``
-    failure. The project is process context (``--project``, ADR-0006).
-    """
-
-    preset: str = Field(
-        description="The export preset's display name, as 'gda export list' reports it."
-    )
-
-
-class ExportGetResult(BaseModel):
-    """The result of ``gda export get``: one preset's details + template readiness (issue #114).
-
-    Echoes the addressed preset's ``index``/``name``/``platform``/``runnable``
-    (read from ``export_presets.cfg``) plus its ``export_path`` (the output path
-    the preset writes to, empty when unset). ``templates_installed`` reports
-    whether the export templates for the running engine version are installed —
-    the readiness check an agent makes before a future ``export run`` (issue
-    #121); ``templates_version`` names the version directory that was checked
-    (e.g. ``4.6.3.stable``), so the agent knows which templates to install when
-    they are missing.
-    """
-
-    index: int = Field(
-        description="The preset's 0-based position in export_presets.cfg (its preset.N section number)."
-    )
-    name: str = Field(description="The preset's display name.")
-    platform: str = Field(
-        description="The preset's target platform (e.g. Linux/X11, Web, macOS)."
-    )
-    runnable: bool = Field(
-        description="Whether the preset is marked runnable (one-click deploy)."
-    )
-    export_path: str = Field(
-        description="The output path the preset exports to, or empty when unset."
-    )
-    templates_installed: bool = Field(
-        description=(
-            "Whether the export templates for the running engine version are "
-            "installed — the readiness check before an export run."
-        )
-    )
-    templates_version: str = Field(
-        description=(
-            "The export-templates version directory checked for installation "
-            "(e.g. 4.6.3.stable), matching the running engine version."
-        )
-    )
-
-
-class ExportRunMode(str, Enum):
-    """The export flavor ``gda export run`` produces (issue #121, selectable #170).
-
-    Maps to Godot's native export flags (ADR-0001). ``release``/``debug`` produce
-    a full platform binary and require the matching export templates to be
-    installed; ``pack`` produces project data only — a PCK/ZIP, chosen by the
-    output path's extension — and needs no platform templates.
-
-    Issue #121 fixed the mode to ``release`` (the common intent — a complete
-    export); follow-up #170 exposes ``--mode`` so an agent can select
-    ``debug``/``pack``. ``release`` stays the default.
-    """
-
-    RELEASE = "release"
-    DEBUG = "debug"
-    PACK = "pack"
-
-
-class ExportRunParams(BaseModel):
-    """The operation params of ``gda export run`` (issue #121, overrides #170).
-
-    ``preset`` addresses the export preset by its display name (as ``export
-    list`` reports it); an unknown name is the ``export_preset_not_found``
-    failure. ``mode`` selects the export flavor (``release`` default; #170).
-    ``output`` overrides the preset's *configured* ``export_path`` (#170); when
-    omitted the export targets the configured path (an empty configured path with
-    no override is the ``export_path_unset`` failure). The project is process
-    context (``--project``, ADR-0006).
-    """
-
-    preset: str = Field(
-        description="The export preset's display name, as 'gda export list' reports it."
-    )
-    mode: ExportRunMode = Field(
-        default=ExportRunMode.RELEASE,
-        description="The export flavor to run (release/debug/pack); default release.",
-    )
-    output: ExportOutputPath | None = Field(
-        default=None,
-        description=(
-            "Override the preset's configured export_path; a relative filesystem "
-            "path is resolved against the invoker's current working directory "
-            "before export."
-        ),
-    )
-
-
-class ExportRunResult(BaseModel):
-    """The result of ``gda export run``: the artifact that was produced (issue #121).
-
-    Echoes the addressed preset's ``preset`` name and target ``platform`` (read
-    from ``export_presets.cfg``), the ``mode`` that was run (the selected flavor,
-    ``release`` by default; #170), and the resolved absolute ``output_path`` the
-    artifact was written to — the effective destination, i.e. the ``--output``
-    override when given, else the preset's configured ``export_path`` resolved
-    against the project directory (#403). ``created_dirs`` lists output parent
-    directories created before the native export, from outermost to innermost
-    (#402).
-    ``warnings`` carries the engine's non-fatal export warnings (e.g. a missing
-    optional icon), parsed best-effort from the export's stderr; an export that
-    succeeds cleanly reports ``warnings == []``. Unlike the sentinel operations,
-    ``export run`` is a native Godot export (the export subsystem is editor-only,
-    ADR-0002 sentinels do not apply), so this result is synthesized by ``gda``
-    from the export's exit code + stderr.
-    """
-
-    preset: str = Field(description="The export preset's display name.")
-    platform: str = Field(
-        description="The preset's target platform (e.g. Linux/X11, Web, macOS)."
-    )
-    mode: ExportRunMode = Field(description="The export flavor that was run.")
-    output_path: str = Field(
-        description="The resolved absolute path the export artifact was written to."
-    )
-    created_dirs: list[str] = Field(
-        description=(
-            "Output parent directories created before export, from outermost to innermost."
-        )
-    )
-    warnings: list[str] = Field(
-        default_factory=list,
-        description="The engine's non-fatal export warnings, parsed from stderr; empty on a clean export.",
-    )
-
-
 class ResourceReference(BaseModel):
     """One referencing site found by ``gda project find-references`` (issue #116).
 
@@ -692,147 +439,9 @@ class ResourceReference(BaseModel):
     )
 
 
-class ProjectFindReferencesResult(BaseModel):
-    """The result of ``gda project find-references``: every site referencing the target.
-
-    Echoes the ``target`` and the list of referencing sites. A target nothing
-    references is a valid, empty result (``references == []``), not a failure —
-    that emptiness is exactly what ``find-unused-resources`` keys on, so the two
-    stay consistent (same reference graph).
-    """
-
-    target: str
-    references: list[ResourceReference]
-
-
-class ProjectDependenciesParams(BaseModel):
-    """The operation params of ``gda project dependencies`` — none (ADR-0004).
-
-    ``dependencies`` maps every scene/resource in the resolved project to the
-    resources it references (its ``[ext_resource]`` entries). The project is
-    process context (``--project``), not an operation param (ADR-0006), so the
-    ``input`` schema is trivially empty.
-    """
-
-
-class Dependency(BaseModel):
-    """One outgoing reference of a scene/resource (issue #116).
-
-    ``path`` is the referenced resource's ``res://`` path; ``kind`` names how it
-    is referenced (``ext_resource`` for a ``[ext_resource]`` entry).
-    """
-
-    path: str = Field(description="The res:// path of the referenced resource.")
-    kind: str = Field(description="How the resource is referenced (ext_resource).")
-
-
-class ResourceDependencies(BaseModel):
-    """One scene/resource and the resources it references (issue #116).
-
-    ``path`` is the scene/resource's own ``res://`` path; ``depends_on`` lists the
-    resources it references. A scene with no external references is reported with
-    an empty ``depends_on``, not dropped.
-    """
-
-    path: str = Field(description="The res:// path of the scene/resource.")
-    depends_on: list[Dependency]
-
-
-class ProjectDependenciesResult(BaseModel):
-    """The result of ``gda project dependencies``: the scene/resource → resource map.
-
-    An empty project is a valid, empty map (``dependencies == []``), not a
-    failure. Built from the same ext_resource parse as ``find-references``, so the
-    two views of the reference graph stay consistent.
-    """
-
-    dependencies: list[ResourceDependencies]
-
-
-class ProjectFindUnusedResourcesParams(BaseModel):
-    """The operation params of ``gda project find-unused-resources`` — none (ADR-0004).
-
-    Reports resource files that nothing references, built on the SAME reference
-    graph as ``find-references``/``dependencies`` (acceptance criterion: the three
-    must agree). The project is process context (``--project``), not an operation
-    param (ADR-0006), so the ``input`` schema is trivially empty.
-    """
-
-
-class ProjectFindUnusedResourcesResult(BaseModel):
-    """The result of ``gda project find-unused-resources``: the unreferenced resources.
-
-    ``unused`` lists the ``res://`` paths of resource files that no other file
-    references AND that are not project entry points (the main scene, an autoload
-    script). A resource is unused exactly when ``find-references`` for it would
-    return an empty list — the consistency the issue requires. An empty list means
-    every resource is referenced (or an entry point).
-    """
-
-    unused: list[str]
-
-
-class ProjectStatisticsParams(BaseModel):
-    """The operation params of ``gda project statistics`` — none (ADR-0004).
-
-    Reports file/line counts, autoloads and plugins for the resolved project. The
-    project is process context (``--project``), not an operation param (ADR-0006),
-    so the ``input`` schema is trivially empty.
-    """
-
-
-class ExtensionCount(BaseModel):
-    """File/line counts for one file extension (issue #116).
-
-    ``extension`` is the lowercased extension without the dot (``gd``, ``tscn``);
-    ``files`` is how many files carry it; ``lines`` is their summed line count.
-    """
-
-    extension: str = Field(
-        description="The lowercased file extension without the dot (e.g. gd, tscn)."
-    )
-    files: int
-    lines: int
-
-
-class Autoload(BaseModel):
-    """One project autoload singleton (issue #116).
-
-    ``name`` is the singleton name; ``path`` its ``res://`` script/scene path
-    (with any leading ``*`` enable marker stripped). Read from ProjectSettings,
-    not by executing the autoload.
-    """
-
-    name: str
-    path: str = Field(
-        description="The autoload's res:// path (enable marker stripped)."
-    )
-
-
-class ProjectStatisticsResult(BaseModel):
-    """The result of ``gda project statistics``: file/line counts, autoloads, plugins.
-
-    ``total_files``/``total_lines`` are the project-wide totals; ``by_extension``
-    breaks them down per extension. ``autoloads`` lists the project's autoload
-    singletons (name + path, read from ProjectSettings); ``plugins`` lists the
-    enabled editor plugins' ``plugin.cfg`` ``res://`` paths. ``scene_count`` /
-    ``script_count`` / ``resource_count`` are convenience counts of ``.tscn`` /
-    ``.gd`` / other-resource files. Line counts cover text files only; binary
-    assets contribute to file counts but not line counts.
-    """
-
-    total_files: int
-    total_lines: int
-    by_extension: list[ExtensionCount]
-    autoloads: list[Autoload]
-    plugins: list[str] = Field(
-        description="The res:// paths of the enabled editor plugins' plugin.cfg files."
-    )
-    scene_count: int
-    script_count: int
-    resource_count: int
-
-
+# Stays in the shared core (ADR-0040 §4): two groups read the SAME engine-version
+# shape — the ``info`` meta command and ``project info`` — so it is a
+# cross-command contract, not one group's model.
 class EngineVersion(BaseModel):
     """The Godot engine version, as reported by ``Engine.get_version_info()``.
 
@@ -849,435 +458,6 @@ class EngineVersion(BaseModel):
     hash: str
     string: str
     timestamp: int
-
-
-class SkillParams(BaseModel):
-    """The operation params of ``gda skill`` (ADR-0024, extended ADR-0027).
-
-    ``gda skill`` is a pure emitter meta command: it reads the bundled
-    ``SKILL.md`` from the package and emits or installs it — no Godot is spawned.
-    ``install`` writes the manifest to a target instead of printing it. The target
-    is named one of two ways: a caller-supplied ``install_dir`` (the neutral path,
-    ADR-0024 — core carries no default location), or a known ``provider`` whose
-    skills directory is resolved at ``scope`` (the opt-in convenience, ADR-0027).
-    The two are mutually exclusive; ``provider`` normalizes to ``install_dir`` here so
-    the argv and ``--params-json`` paths resolve identically (ADR-0015).
-    """
-
-    install: bool = Field(
-        default=False,
-        description="If true, WRITE the bundled SKILL.md to the target "
-        "instead of returning it; the result then reports the written path.",
-    )
-    install_dir: str | None = Field(
-        default=None,
-        description="The skills directory to install into (caller-supplied; the neutral "
-        "path, no default). Parent dirs are created and an existing file is overwritten. "
-        "Providing it implies an install (ADR-0015 parity with argv --dir). Mutually "
-        "exclusive with provider.",
-    )
-    provider: SkillProvider | None = Field(
-        default=None,
-        description="Install into a KNOWN agent's skills directory instead of a "
-        "caller-supplied install_dir: resolves that agent's directory at scope "
-        "(ADR-0027). Mutually exclusive with install_dir; providing it implies an install.",
-    )
-    scope: SkillScope = Field(
-        default=SkillScope.USER,
-        description="With provider, whether to install into the agent's per-project "
-        "(committed) or per-user (all projects) skills directory; default user.",
-    )
-
-    @model_validator(mode="after")
-    def _resolve_install_target(self) -> "SkillParams":
-        # Single source of truth (ADR-0015): normalize the target HERE, in the model, so
-        # argv and a --params-json object agree. A named provider resolves to its known
-        # skills dir (ADR-0027) — but provider and an explicit install_dir name the SAME
-        # thing two ways, so giving both is ambiguous and rejected. Then, whichever way a
-        # target was named, naming one means "install there".
-        if self.provider is not None:
-            if self.install_dir is not None:
-                raise ValueError(
-                    "provider and install_dir are mutually exclusive: name an agent "
-                    "(--provider) OR a directory (--dir), not both"
-                )
-            self.install_dir = resolve_skill_dir(self.provider, self.scope)
-        if self.install_dir is not None:
-            self.install = True
-        return self
-
-
-class SkillResult(BaseModel):
-    """The result of ``gda skill``: the bundled Skill, version-locked (ADR-0024).
-
-    ``name``/``version``/``content`` carry the manifest's identity, the installed
-    ``gda`` version (from ``importlib.metadata``, so the guidance cannot skew from
-    the CLI it describes), and the full ``SKILL.md`` text. ``installed_path`` is the
-    path written on ``--install`` and ``None`` for a plain emit, so one model serves
-    both the emit and install paths.
-    """
-
-    name: str
-    version: str
-    content: str
-    installed_path: Path | None = None
-
-
-class ProjectInfoParams(BaseModel):
-    """The operation params of ``gda project info`` — none (ADR-0004).
-
-    ``project info`` reports the resolved project's metadata from its
-    ``ProjectSettings``; the project is process context (``--project``), not an
-    operation param (ADR-0006), so the ``input`` schema is trivially empty.
-    """
-
-
-class ProjectInfoResult(BaseModel):
-    """The result of ``gda project info``: core project metadata (issue #111).
-
-    Reports the project's ``name`` and ``main_scene`` from ``ProjectSettings``,
-    its configured ``viewport_width``/``viewport_height``, and the ``engine_version``
-    the project runs on (the same shape ``gda info`` reports). ``main_scene`` is the
-    empty string for a project that has not set a main scene, and the viewport
-    fields fall back to the engine's built-in defaults when the project never set
-    them, so a brand-new project still reports a complete, valid result.
-    """
-
-    name: str = Field(
-        description="The project name (ProjectSettings application/config/name)."
-    )
-    main_scene: str = Field(
-        description=(
-            "The project's main scene path (application/run/main_scene), or the "
-            "empty string when none is set."
-        )
-    )
-    viewport_width: int = Field(
-        description="The configured viewport width (display/window/size/viewport_width)."
-    )
-    viewport_height: int = Field(
-        description="The configured viewport height (display/window/size/viewport_height)."
-    )
-    engine_version: EngineVersion = Field(
-        description="The Godot engine version the project runs on."
-    )
-
-
-class ProjectGetParams(BaseModel):
-    """The operation params of ``gda project get`` (issue #111).
-
-    ``setting`` is the project setting's full ``section/key`` name (e.g.
-    ``application/config/name``). The project is process context (``--project``),
-    not an operation param (ADR-0006), so only ``setting`` is an input.
-    """
-
-    setting: str = Field(
-        description=(
-            "The project setting's full section/key name, e.g. application/config/name."
-        )
-    )
-
-
-class ProjectGetResult(BaseModel):
-    """The result of ``gda project get``: one setting as typed JSON (issue #111).
-
-    Echoes the addressed ``setting``, its declared Godot ``type`` name, and its
-    ``value`` in the same JSON projection ``node get`` reports for a node property
-    (a scalar stays a scalar, a Vector2 becomes ``[x, y]``) — the projection
-    ``project set`` accepts back, so a ``set`` round-trips through a ``get``.
-    """
-
-    setting: str
-    type: str = Field(
-        description="The setting's declared Godot type name (e.g. String, int, Vector2)."
-    )
-    value: Any = Field(
-        description="The setting's value as JSON. " + _VALUE_PROJECTION_DESC,
-        json_schema_extra=_projected_value_schema_extra,
-    )
-
-
-class ProjectListParams(BaseModel):
-    """The operation params of ``gda project list`` (issue #312).
-
-    Enumerates the project's ``ProjectSettings`` keys so an agent can DISCOVER
-    which settings exist — the list half of the ``list → get → set`` workflow
-    (``get``/``set`` both require you to already know the ``section/key``).
-    ``include_defaults`` (the CLI ``--all`` flag) widens the listing from only the
-    project's customized (non-default) settings to the engine's built-in defaults
-    too; ``section`` restricts it to keys whose name begins with that ``section/``
-    prefix (e.g. ``application/``, ``display/``), and the two compose. The project
-    is process context (``--project``), not an operation param (ADR-0006).
-    """
-
-    include_defaults: bool = Field(
-        default=False,
-        description=(
-            "Also list the engine's built-in default settings, not just the "
-            "project's customized (non-default) ones. The CLI --all flag."
-        ),
-    )
-    section: str | None = Field(
-        default=None,
-        description=(
-            "Restrict the listing to keys whose name begins with this section/ "
-            "prefix, e.g. application/ or display/. Null lists every section."
-        ),
-    )
-
-
-class ListedProjectSetting(BaseModel):
-    """One enumerated project setting of ``gda project list`` (issue #312).
-
-    Reuses the same ``{setting, type, value}`` projection ``project get`` reports
-    for a single setting — so a listed entry round-trips through ``project get`` —
-    plus ``is_default``: ``false`` when the key is customized (written in
-    ``project.godot``), ``true`` when it is at the engine's built-in default.
-    """
-
-    setting: str
-    type: str = Field(
-        description="The setting's declared Godot type name (e.g. String, int, Vector2)."
-    )
-    value: Any = Field(
-        description="The setting's value as JSON. " + _VALUE_PROJECTION_DESC,
-        json_schema_extra=_projected_value_schema_extra,
-    )
-    is_default: bool = Field(
-        description=(
-            "True when the setting is at the engine's built-in default; false when "
-            "it is customized in project.godot."
-        )
-    )
-
-
-class ProjectListResult(BaseModel):
-    """The result of ``gda project list``: the project's enumerated settings (#312).
-
-    ``settings`` is the discovered keys, each a ``{setting, type, value,
-    is_default}`` entry sorted by name. A project with no customized settings and
-    no ``include_defaults`` is a valid, EMPTY listing — ``settings == []`` — not a
-    failure. Internal engine-bookkeeping and non-setting properties are filtered
-    out, so only real ``ProjectSettings`` keys appear.
-    """
-
-    settings: list[ListedProjectSetting]
-
-
-class ProjectSetParams(BaseModel):
-    """The operation params of ``gda project set`` (issue #111).
-
-    ``setting`` is the project setting's full ``section/key`` name; ``value`` is
-    the CLI string value, coerced to the setting's declared Godot type by the
-    operation (the same coercion rules as ``node set``, #55) before
-    ``project.godot`` is saved. ``set`` edits an EXISTING setting — an unknown key
-    is a clean error, never a silent create — so the declared type to coerce to is
-    always known (read off the setting's current value).
-    """
-
-    setting: str = Field(
-        description=(
-            "The project setting's full section/key name, e.g. application/config/name."
-        )
-    )
-    value: str = Field(
-        description=(
-            "The value to set, as a string. The operation coerces it to the "
-            "setting's declared Godot type (see the command catalog's 'Property "
-            "value coercion'). For Dictionary/Array JSON values, JSON integer "
-            "literals stay int and JSON float literals stay float; typed "
-            "containers assign entries through their declared container type. An "
-            "uncoercible value is a clean error."
-        )
-    )
-
-
-class ProjectSetResult(BaseModel):
-    """The result of ``gda project set``: the one setting it set (issue #111).
-
-    Echoes the ``setting`` set, the declared ``type`` the CLI value was coerced
-    to, and the coerced ``value`` as JSON — the same projection ``project get``
-    reports, so a ``set`` round-trips through a ``get`` without re-reading
-    ``project.godot``.
-    """
-
-    setting: str
-    type: str = Field(
-        description="The setting's declared Godot type the value was coerced to."
-    )
-    value: Any = Field(
-        description=(
-            "The coerced value as JSON, as ProjectSettings now holds it. "
-            + _SET_ECHO_VALUE_DESC
-        ),
-        json_schema_extra=_projected_value_schema_extra,
-    )
-
-
-class ProjectAddAutoloadParams(BaseModel):
-    """The operation params of ``gda project add-autoload`` (issue #119).
-
-    Registers an autoload singleton: ``name`` is the global accessor name (the key
-    under the ``autoload/`` section of ``project.godot``); ``path`` is the res://
-    path to the script or scene to autoload. The operation stores it in the
-    enabled-singleton form (a leading ``*`` prefix) and saves ``project.godot``.
-    The project is process context (``--project``), not an operation param
-    (ADR-0006), so only ``name`` and ``path`` are inputs.
-    """
-
-    name: str = Field(
-        description=(
-            "The autoload singleton's global name — the accessor it is reached by "
-            "and the key under the project's autoload/ section."
-        )
-    )
-    path: NormalizedPath = Field(
-        description=(
-            "The res:// path to the script or scene to autoload, e.g. res://global.gd."
-        )
-    )
-
-
-class ProjectAddAutoloadResult(BaseModel):
-    """The result of ``gda project add-autoload``: the autoload it registered.
-
-    Echoes the autoload's ``name`` and the ``path`` exactly as it was persisted to
-    ``project.godot`` — the enabled-singleton form with the leading ``*`` prefix
-    (e.g. ``*res://global.gd``), the same value a ``project get`` of
-    ``autoload/<name>`` reads back, so an add round-trips through a get.
-    """
-
-    name: str = Field(description="The registered autoload's global name.")
-    path: str = Field(
-        description=(
-            "The autoload value as persisted to project.godot, in enabled-singleton "
-            "form with the leading * prefix (e.g. *res://global.gd)."
-        )
-    )
-
-
-class ProjectRemoveAutoloadParams(BaseModel):
-    """The operation params of ``gda project remove-autoload`` (issue #119).
-
-    Unregisters an autoload singleton by its global ``name`` (the key under the
-    ``autoload/`` section), then saves ``project.godot``. The project is process
-    context (``--project``), not an operation param (ADR-0006), so only ``name``
-    is an input.
-    """
-
-    name: str = Field(
-        description="The global name of the autoload singleton to unregister."
-    )
-
-
-class ProjectRemoveAutoloadResult(BaseModel):
-    """The result of ``gda project remove-autoload``: the autoload it removed.
-
-    Echoes the ``name`` of the autoload that was unregistered, so an agent can
-    confirm which singleton was removed; a subsequent ``project get`` of
-    ``autoload/<name>`` reports ``unknown_setting``.
-    """
-
-    name: str = Field(description="The unregistered autoload's global name.")
-
-
-class ProjectAddInputActionParams(BaseModel):
-    """The operation params of ``gda project add-input-action`` (issue #380).
-
-    Registers an InputMap action: ``name`` is the action name (the key under the
-    ``input/`` section of ``project.godot``), bound to one or more keyboard keys.
-    The operation builds real ``InputEventKey`` events and persists the action
-    via ``ProjectSettings`` — never a hand-built string — so the serialization is
-    exactly the engine's own ``var_to_str`` form. The project is process context
-    (``--project``), not an operation param (ADR-0006).
-    """
-
-    name: str = Field(
-        description=(
-            "The input action's name — the key under the project's input/ section "
-            "and the name gda input action drives it by."
-        )
-    )
-    keys: list[str] = Field(
-        min_length=1,
-        description=(
-            "The keys to bind (at least one): each item is a Godot key NAME "
-            "(e.g. J, Space, Escape) or a base-10 keycode integer string."
-        ),
-    )
-    deadzone: float = Field(
-        default=0.5,
-        ge=0.0,
-        le=1.0,
-        description=(
-            "The action's deadzone, 0..1 (the editor slider's bounds); "
-            "Godot's default is 0.5."
-        ),
-    )
-    physical: bool = Field(
-        default=False,
-        description=(
-            "Bind physical_keycode (keyboard position, layout-independent) "
-            "instead of keycode."
-        ),
-    )
-
-
-class InputActionKeyEvent(BaseModel):
-    """One key binding of a registered InputMap action (issue #380).
-
-    ``kind`` discriminates the event type so mouse/joypad kinds can extend the
-    shape later without breaking; this slice emits only ``key`` events. ``key``
-    echoes the raw ``--key`` token, ``keycode`` the Godot keycode it resolved to,
-    and ``physical`` whether it was bound as ``physical_keycode``.
-    """
-
-    kind: str = Field(
-        default="key", description="The event kind ('key' for this slice)."
-    )
-    key: str = Field(description="The raw --key token as given (name or keycode).")
-    keycode: int = Field(description="The Godot keycode the token resolved to.")
-    physical: bool = Field(
-        description="True when bound as physical_keycode instead of keycode."
-    )
-
-
-class ProjectAddInputActionResult(BaseModel):
-    """The result of ``gda project add-input-action``: the action it registered.
-
-    Echoes the action's ``name``, the ``deadzone`` persisted, and the resolved
-    key ``events`` exactly as they were bound — so an agent can confirm each key
-    token mapped to the intended keycode without re-reading ``project.godot``.
-    """
-
-    name: str = Field(description="The registered input action's name.")
-    deadzone: float = Field(description="The deadzone persisted with the action.")
-    events: list[InputActionKeyEvent] = Field(
-        description="The key events bound to the action, in --key order."
-    )
-
-
-class ProjectRemoveInputActionParams(BaseModel):
-    """The operation params of ``gda project remove-input-action`` (issue #380).
-
-    Unregisters an InputMap action by its ``name`` (the key under the ``input/``
-    section), then saves ``project.godot``. The project is process context
-    (``--project``), not an operation param (ADR-0006), so only ``name`` is an
-    input.
-    """
-
-    name: str = Field(description="The name of the input action to unregister.")
-
-
-class ProjectRemoveInputActionResult(BaseModel):
-    """The result of ``gda project remove-input-action``: the action it removed.
-
-    Echoes the ``name`` of the input action that was unregistered, so an agent
-    can confirm which action was removed; a subsequent ``project get`` of
-    ``input/<name>`` reports ``unknown_setting``.
-    """
-
-    name: str = Field(description="The unregistered input action's name.")
 
 
 class GameNode(BaseModel):
