+++ src/gda/render.py
@@ -37,10 +37,6 @@ if TYPE_CHECKING:
         DaemonStopResult,
         DaemonUninstallResult,
         DiagErrorsResult,
-        EngineVersion,
-        ExportGetResult,
-        ExportListResult,
-        ExportRunResult,
         GameGetResult,
         GameNode,
         GameRectResult,
@@ -53,21 +49,8 @@ if TYPE_CHECKING:
         LoggerTailResult,
         PerfMonitorResult,
         PerfMonitorsResult,
-        ProjectAddAutoloadResult,
-        ProjectAddInputActionResult,
-        ProjectGetResult,
-        ProjectInfoResult,
-        ProjectListResult,
-        ProjectRemoveAutoloadResult,
-        ProjectRemoveInputActionResult,
-        ProjectSetResult,
         ScreenCaptureResult,
         ScreenFramesResult,
-        SkillResult,
-        ProjectDependenciesResult,
-        ProjectFindReferencesResult,
-        ProjectFindUnusedResourcesResult,
-        ProjectStatisticsResult,
     )
 
 
@@ -319,179 +302,3 @@ def render_daemon_uninstall(uninstalled: "DaemonUninstallResult") -> str:
     if uninstalled.removed:
         return "harness uninstalled"
     return "no harness was installed"
-
-
-def render_export_list(listed: "ExportListResult") -> str:
-    """Render the enumerated presets as ``name (platform) [runnable]`` lines."""
-    if not listed.presets:
-        return "(no presets)"
-    lines = []
-    for preset in listed.presets:
-        runnable = " [runnable]" if preset.runnable else ""
-        lines.append(f"{preset.name} ({preset.platform}){runnable}")
-    return "\n".join(lines)
-
-
-def render_export_get(got: "ExportGetResult") -> str:
-    """Render one preset's details plus its export-template readiness."""
-    runnable = " [runnable]" if got.runnable else ""
-    header = f"{got.name} ({got.platform}){runnable}"
-    templates = (
-        f"templates installed ({got.templates_version})"
-        if got.templates_installed
-        else f"templates missing ({got.templates_version})"
-    )
-    return "\n".join([header, f"  export_path: {got.export_path}", f"  {templates}"])
-
-
-def render_export_run(ran: "ExportRunResult") -> str:
-    """Render a completed export as ``exported <preset> (<platform>, <mode>) -> <path>``.
-
-    Echoes the artifact that was produced, then one ``warning: …`` line per
-    non-fatal engine warning (a clean export prints just the header line).
-    """
-    header = (
-        f"exported {ran.preset} ({ran.platform}, {ran.mode.value}) -> {ran.output_path}"
-    )
-    if not ran.warnings:
-        return header
-    return "\n".join([header, *[f"  warning: {w}" for w in ran.warnings]])
-
-
-def render_project_info(info: "ProjectInfoResult") -> str:
-    """Render project metadata as a small ``key: value`` block for humans."""
-    main_scene = info.main_scene if info.main_scene else "(none)"
-    return "\n".join(
-        [
-            f"name: {info.name}",
-            f"main_scene: {main_scene}",
-            f"viewport: {info.viewport_width}x{info.viewport_height}",
-            f"engine: {info.engine_version.string}",
-        ]
-    )
-
-
-def render_project_get(got: "ProjectGetResult") -> str:
-    """Render a read setting as ``<setting> (<type>) = <value>``."""
-    return f"{got.setting} ({got.type}) = {format_value(got.value)}"
-
-
-def render_project_set(was_set: "ProjectSetResult") -> str:
-    """Render a set setting as ``set <setting> (<type>) = <value>``."""
-    return f"set {was_set.setting} ({was_set.type}) = {format_value(was_set.value)}"
-
-
-def render_project_list(listed: "ProjectListResult") -> str:
-    """Render enumerated settings as ``<setting> (<type>) = <value>`` lines.
-
-    An engine-default entry is tagged ``[default]`` so customized vs default reads
-    at a glance; the same ``<setting> (<type>) = <value>`` shape ``project get``
-    renders. An empty listing is named explicitly rather than printing nothing.
-    """
-    if not listed.settings:
-        return "(no settings)"
-    lines = []
-    for entry in listed.settings:
-        default_marker = " [default]" if entry.is_default else ""
-        lines.append(
-            f"{entry.setting} ({entry.type}) = {format_value(entry.value)}"
-            f"{default_marker}"
-        )
-    return "\n".join(lines)
-
-
-def render_project_add_autoload(added: "ProjectAddAutoloadResult") -> str:
-    """Render a registered autoload as ``added autoload <name> = <path>``."""
-    return f"added autoload {added.name} = {added.path}"
-
-
-def render_project_remove_autoload(removed: "ProjectRemoveAutoloadResult") -> str:
-    """Render an unregistered autoload as ``removed autoload <name>``."""
-    return f"removed autoload {removed.name}"
-
-
-def render_project_add_input_action(added: "ProjectAddInputActionResult") -> str:
-    """Render a registered input action with its resolved key bindings.
-
-    e.g. ``added input action jump (deadzone 0.5): J -> 74, Space -> 32``; a
-    physical binding is marked ``(physical)`` after its keycode.
-    """
-    bindings = ", ".join(
-        f"{event.key} -> {event.keycode}" + (" (physical)" if event.physical else "")
-        for event in added.events
-    )
-    return f"added input action {added.name} (deadzone {added.deadzone}): {bindings}"
-
-
-def render_project_remove_input_action(
-    removed: "ProjectRemoveInputActionResult",
-) -> str:
-    """Render an unregistered input action as ``removed input action <name>``."""
-    return f"removed input action {removed.name}"
-
-
-def render_project_find_references(found: "ProjectFindReferencesResult") -> str:
-    """Render find-references as ``<target>`` then one ``path (kind)`` line each."""
-    if not found.references:
-        return f"{found.target} (no references)"
-    lines = [found.target]
-    lines += [f"  {ref.path} ({ref.kind})" for ref in found.references]
-    return "\n".join(lines)
-
-
-def render_project_dependencies(deps: "ProjectDependenciesResult") -> str:
-    """Render the dependency map as ``<scene>`` then indented ``-> <dep>`` lines."""
-    if not deps.dependencies:
-        return "(no scenes or resources)"
-    lines = []
-    for resource in deps.dependencies:
-        lines.append(resource.path)
-        lines += [f"  -> {dep.path} ({dep.kind})" for dep in resource.depends_on]
-    return "\n".join(lines)
-
-
-def render_project_find_unused_resources(
-    unused: "ProjectFindUnusedResourcesResult",
-) -> str:
-    """Render the unreferenced resources, one path per line."""
-    if not unused.unused:
-        return "(no unused resources)"
-    return "\n".join(unused.unused)
-
-
-def render_project_statistics(stats: "ProjectStatisticsResult") -> str:
-    """Render the project statistics as a human-readable summary."""
-    lines = [
-        f"{stats.total_files} files, {stats.total_lines} lines",
-        (
-            f"  scenes: {stats.scene_count}, scripts: {stats.script_count}, "
-            f"resources: {stats.resource_count}"
-        ),
-    ]
-    for ext in stats.by_extension:
-        lines.append(f"  .{ext.extension}: {ext.files} files, {ext.lines} lines")
-    if stats.autoloads:
-        lines.append("autoloads:")
-        lines += [f"  {a.name} = {a.path}" for a in stats.autoloads]
-    if stats.plugins:
-        lines.append("plugins:")
-        lines += [f"  {p}" for p in stats.plugins]
-    return "\n".join(lines)
-
-
-def render_engine_version(version: "EngineVersion") -> str:
-    """Render the engine version as its one-line version string."""
-    return version.string
-
-
-def render_skill(skill: "SkillResult") -> str:
-    """Render ``gda skill`` as text (ADR-0024).
-
-    A plain emit prints the raw ``SKILL.md`` verbatim, so
-    ``gda skill > .../SKILL.md`` drops the manifest straight to disk; an
-    ``--install`` instead reports the written path (the file already holds the
-    same content) rather than echoing it twice.
-    """
-    if skill.installed_path is not None:
-        return f"Installed the gda Skill to {skill.installed_path}"
-    return skill.content
