+++ tests/test_classify_export_run.py
@@ -12,14 +12,14 @@ pipeline.
 
 from pathlib import Path
 
-from gda.errors import (
-    Failure,
+from gda.commands.export import (
+    ExportRunMode,
+    ExportRunResult,
     classify_export_run,
-    export_path_unset_failure,
     parse_export_warnings,
 )
+from gda.errors import Failure, export_path_unset_failure
 from gda.exit_codes import EXIT_OPERATION
-from gda.models import ExportRunMode, ExportRunResult
 from gda.runner import LaunchFailure, RunResult
 
 BINARY = Path("/x/Godot")
