+++ tests/test_models.py
@@ -37,18 +37,14 @@ from gda.commands.script import (
     ScriptSetResult,
     ScriptValidateResult,
 )
-from gda.models import (
-    EngineVersion,
+from gda.commands.export import (
     ExportGetResult,
     ExportListResult,
     ExportRunMode,
     ExportRunResult,
-    GameSetResult,
-    GdaError,
-    GdaErrorEnvelope,
-    InlineValueProjection,
+)
+from gda.commands.project import (
     ListedProjectSetting,
-    NodeProperty,
     ProjectAddAutoloadResult,
     ProjectAddInputActionResult,
     ProjectGetResult,
@@ -56,6 +52,14 @@ from gda.models import (
     ProjectRemoveAutoloadResult,
     ProjectRemoveInputActionResult,
     ProjectSetResult,
+)
+from gda.models import (
+    EngineVersion,
+    GameSetResult,
+    GdaError,
+    GdaErrorEnvelope,
+    InlineValueProjection,
+    NodeProperty,
     ReferenceProjection,
 )
